Fix for multiple module issue
This commit fixes the lingering handlers from creating a new module. This was fixed by allowing the garbage collector to collect handlers by using a WeakDelegate instead of a regular one for the MessageBroker.

Also fixed an issue with deserializaton of Equipment, the previous Package factory was keeping the previous equipment alive. Made package factory transient to fix