Implementation of a ArmyQueueManager
When the army is not create yet in the client instance, because message throw the network don't arrive in the same order, this mobile party is add in a queue. Later, for now 1 second, the queue will be fetch by the manager to try to set the mobile party to the army again. Works great for now.
I'll change first the way mobile party add are set, by setting the list directly in the game, to suppress the duplication issue